完成基础调用demo

+++ model/model.go
@@ -17,7 +17,7 @@ import (
 
 
 type JsonResult struct {
-	Code uint8       `json:"code"`
+	Code int       `json:"code"`
 	Msg  string      `json:"msg"`
 	Data interface{} `json:"data,omitempty"`
 }
